flaskapp/requirements.txt
# packages in environment at C:\Users\timty\anaconda3\envs\soft_eng_project:
#
# Name                    Version                   Build  Channel
argon2-cffi               20.1.0                   pypi_0    pypi
async-generator           1.10                     pypi_0    pypi
attrs                     20.3.0                   pypi_0    pypi
backcall                  0.2.0                    pypi_0    pypi
bleach                    3.3.0                    pypi_0    pypi
ca-certificates           2021.1.19            haa95532_0  
certifi                   2020.12.5        py38haa95532_0  
cffi                      1.14.5                   pypi_0    pypi
chardet                   4.0.0                    pypi_0    pypi
decorator                 4.4.2                    pypi_0    pypi
defusedxml                0.6.0                    pypi_0    pypi
entrypoints               0.3                      pypi_0    pypi
idna                      2.10                     pypi_0    pypi
ipykernel                 5.4.3                    pypi_0    pypi
ipython                   7.20.0                   pypi_0    pypi
ipython-genutils          0.2.0                    pypi_0    pypi
ipywidgets                7.6.3                    pypi_0    pypi
jedi                      0.18.0           py38haa95532_1  
jinja2                    2.11.3                   pypi_0    pypi
jsonschema                3.2.0                    pypi_0    pypi
jupyter                   1.0.0                    pypi_0    pypi
jupyter-client            6.1.11                   pypi_0    pypi
jupyter-console           6.2.0                    pypi_0    pypi
jupyter-core              4.7.1                    pypi_0    pypi
jupyterlab-pygments       0.1.2                    pypi_0    pypi
jupyterlab-widgets        1.0.0                    pypi_0    pypi
markupsafe                1.1.1                    pypi_0    pypi
mistune                   0.8.4                    pypi_0    pypi
mysql                     0.0.2                    pypi_0    pypi
mysql-connector-python    8.0.23                   pypi_0    pypi
mysql-connector-python-rf 2.2.2                    pypi_0    pypi
mysqlclient               2.0.3                    pypi_0    pypi
nbclient                  0.5.2                    pypi_0    pypi
nbconvert                 6.0.7                    pypi_0    pypi
nbformat                  5.1.2                    pypi_0    pypi
nest-asyncio              1.5.1                    pypi_0    pypi
notebook                  6.2.0                    pypi_0    pypi
numpy                     1.20.1                   pypi_0    pypi
openssl                   1.1.1i               h2bbff1b_0  
packaging                 20.9                     pypi_0    pypi
pandas                    1.2.2                    pypi_0    pypi
pandocfilters             1.4.3                    pypi_0    pypi
parso                     0.8.1              pyhd3eb1b0_0  
pickleshare               0.7.5                    pypi_0    pypi
pip                       20.3.3           py38haa95532_0  
prometheus-client         0.9.0                    pypi_0    pypi
prompt-toolkit            3.0.16                   pypi_0    pypi
protobuf                  3.14.0                   pypi_0    pypi
pycparser                 2.20                     pypi_0    pypi
pygments                  2.8.0                    pypi_0    pypi
pyparsing                 2.4.7                    pypi_0    pypi
pyrsistent                0.17.3                   pypi_0    pypi
python                    3.8.5                h5fd99cc_1  
python-dateutil           2.8.1                    pypi_0    pypi
pytz                      2021.1                   pypi_0    pypi
pywin32                   300                      pypi_0    pypi
pywinpty                  0.5.7                    pypi_0    pypi
pyzmq                     22.0.3                   pypi_0    pypi
qtconsole                 5.0.2                    pypi_0    pypi
qtpy                      1.9.0                    pypi_0    pypi
requests                  2.25.1                   pypi_0    pypi
send2trash                1.5.0                    pypi_0    pypi
setuptools                52.0.0           py38haa95532_0  
simplejson                3.17.2                   pypi_0    pypi
six                       1.15.0                   pypi_0    pypi
sqlalchemy                1.3.23                   pypi_0    pypi
sqlite                    3.33.0               h2a8f88b_0  
terminado                 0.9.2                    pypi_0    pypi
testpath                  0.4.4                    pypi_0    pypi
tornado                   6.1                      pypi_0    pypi
traitlets                 5.0.5                    pypi_0    pypi
urllib3                   1.26.3                   pypi_0    pypi
vc                        14.2                 h21ff451_1  
vs2015_runtime            14.27.29016          h5e58377_2  
wcwidth                   0.2.5                    pypi_0    pypi
webencodings              0.5.1                    pypi_0    pypi
wheel                     0.36.2             pyhd3eb1b0_0  
widgetsnbextension        3.5.1                    pypi_0    pypi
wincertstore              0.2                      py38_0  
zlib                      1.2.11               h62dcd97_4  
